+++ crates/ruma-events/tests/it/file.rs
@@ -12,7 +12,7 @@ use ruma_events::{
     AnyMessageLikeEvent, MessageLikeEvent,
     file::{EncryptedContentInit, FileEventContent},
     message::TextContentBlock,
-    relation::InReplyTo,
+    relation::Reply,
     room::{JsonWebKeyInit, message::Relation},
 };
 use serde_json::{from_value as from_json_value, json};
@@ -100,9 +100,8 @@ fn file_event_serialization() {
     );
     content.file.mimetype = Some("text/plain".to_owned());
     content.file.size = Some(uint!(774));
-    content.relates_to = Some(Relation::Reply {
-        in_reply_to: InReplyTo::new(owned_event_id!("$replyevent:example.com")),
-    });
+    content.relates_to =
+        Some(Relation::Reply(Reply::with_event_id(owned_event_id!("$replyevent:example.com"))));
 
     assert_to_canonical_json_eq!(
         content,