+++ crates/ruma-events/tests/it/video.rs
@@ -15,7 +15,7 @@ use ruma_events::{
     file::{CaptionContentBlock, EncryptedContentInit, FileContentBlock},
     image::{Thumbnail, ThumbnailFileContentBlock, ThumbnailImageDetailsContentBlock},
     message::TextContentBlock,
-    relation::InReplyTo,
+    relation::Reply,
     room::{JsonWebKeyInit, message::Relation},
     video::{VideoDetailsContentBlock, VideoEventContent},
 };
@@ -127,9 +127,8 @@ fn event_serialization() {
     thumbnail.file.size = Some(uint!(334_593));
     content.thumbnail = vec![thumbnail].into();
     content.caption = Some(CaptionContentBlock::plain("This is my awesome vintage lava lamp"));
-    content.relates_to = Some(Relation::Reply {
-        in_reply_to: InReplyTo::new(owned_event_id!("$replyevent:example.com")),
-    });
+    content.relates_to =
+        Some(Relation::Reply(Reply::with_event_id(owned_event_id!("$replyevent:example.com"))));
 
     assert_to_canonical_json_eq!(
         content,